requirements.txt
astroid>=2.4.2
certifi>=2020.6.20
chardet>=3.0.4
click>=7.1.2
colorama>=0.4.3
Flask>=1.1.2
idna>=2.10
isort>=4.3.21
itsdangerous>=1.1.0
# Jinja2>=2.11.2
# lazy-object-proxy>=1.4.3
# MarkupSafe>=1.1.1
# mccabe>=0.6.1
mongoengine>=0.20.0
# mysql-connector>=2.2.9
# mysql-connector-python>=8.0.20
ordereddict>=1.1
protobuf>=3.12.2
pyaes>=1.6.1
pyasn1>=0.4.8
# pylint>=2.5.3
pymongo>=3.10.1
# PyMySQL>=0.9.3
pyTelegramBotAPI>=3.7.2
python-dotenv>=0.13.0
requests>=2.24.0
rsa>=4.1
six>=1.15.0
# SQLAlchemy>=1.3.17
Telethon>=1.14.0
toml>=0.10.1
urllib3>=1.25.9
Werkzeug>=1.0.1
WooCommerce>=2.1.1
wrapt>=1.12.1
apscheduler